feat: move server parts to `parts` folder
